Rename 'Exception' to 'Error' for consistency with Python and C++ usage (#23)
However, note that we keep using 'exceptions.h' as the header where the exceptions are defined,
instead of, say, errors.h, despite all exception names ending with 'Error'.
This also seems the most common practice in Python, at least, and is more generic
for when we'd like to define exceptions which are not errors per se.