Remove comment: decision is to keep the name StringId.
Alternatives where Name, Symbol, or InternedString like in Rust. StringId was chosen
as a good compromise between being explicit (it represents a string) and conciseness (less
to type than InternedString which would be even more explicit). Also, it's nice that the
class name starts with String, instead of ending with String. This is consistent with
StringRef, another class which we plan to implement.